Add files via upload
This is a json based Inventory Management System project which adds new stationery products using the Adding_new_products.ipynb file and keeps the record of the updated inventory in the record.json. Various functions such as updation of the quantity and the price of the existing products, deletion of the existing products along with addition of the new products have been added to the file. The Purchasing_products.ipynb is meant for the end-user to purchase the products. Bills are generated and the total quantity of the product sold along with the amount is updated on the sales.json file. Similarly, the records(quantity) in the record.json file are updated accordingly.